created transactions folder to reuse cypher statements. Put isValidInviteCode resolving transaction there
